requirements.txt
python-telegram-bot==21.0
openai==1.35.0
httpx==0.27.2
woocommerce==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.4.0
